pyproject.toml
[project]
name = "indian-stock-prescriptive-model"
version = "0.1.0"
description = "Trend-based stock prediction for NSE with WhatsApp alerts"
dependencies = [
    "niftystocks",
    "yfinance>=0.2.0",
    "prophet>=1.1.0",
    "xgboost>=2.0.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas-market-calendars>=4.4.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"